backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "sqlmodel>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic[email]>=2.11.7",
]
